# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a Audit log swagger server.",
        "title": "Audit log Swagger API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/tenants": {
            "get": {
                "description": "Get a list of all tenants that the authenticated user has access to",
                "tags": [
                    "tenants"
                ],
                "summary": "List all tenants",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/domain.Tenant"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tenant with specified configuration",
                "tags": [
                    "tenants"
                ],
                "summary": "Create a new tenant",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/domain.Tenant"
                            }
                        }
                    },
                    "description": "Tenant object",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.Tenant"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "servers": [
        {
            "url": "//localhost:10000/api/v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "domain.Tenant": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "rate_limit": {
                        "type": "integer"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "dto.Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "error message"
                    }
                }
            }
        }
    }
}